+++ database/migrations/2015_07_25_151103_create_weather_main_table.php
@@ -29,7 +29,7 @@ class CreateWeatherMainTable extends Migration
         \Schema::create($this->table, function(Blueprint $t){
             
             $t->bigIncrements('id');
-            $t->integer('mainable_id')->unsigned();
+            $t->bigInteger('mainable_id')->unsigned();
             $t->string('mainable_type');          
             $t->float('temp')->nullable();
             $t->float('temp_min')->nullable();
